Make coordinate computations compatible with I>16.

+++ sieve/las-coordinates.h
@@ -5,21 +5,21 @@
 #include "las-types.h"
 
 /*  Forward declarations of conversion functions */
-void xToIJ(int *i, unsigned int *j, const unsigned int X, sieve_info_srcptr si);
+void xToIJ(int *i, unsigned int *j, const uint64_t X, sieve_info_srcptr si);
 void NxToIJ(int *i, unsigned int *j, const unsigned int N, const unsigned int x, sieve_info_srcptr si);
 
-void IJTox(unsigned int * x, int i, unsigned int j, sieve_info_srcptr si);
+void IJTox(uint64_t * x, int i, unsigned int j, sieve_info_srcptr si);
 void IJToNx(unsigned int *N, unsigned int * x, int i, unsigned int j, sieve_info_srcptr si);
 void IJToAB(int64_t *a, uint64_t *b, const int i, const unsigned int j, sieve_info_srcptr si);
-static inline void xToAB(int64_t *a, uint64_t *b, const unsigned int x, sieve_info_srcptr si);
+static inline void xToAB(int64_t *a, uint64_t *b, const uint64_t x, sieve_info_srcptr si);
 static inline void NxToAB(int64_t *a, uint64_t *b, const unsigned int N, const unsigned int x, sieve_info_srcptr si);
 int ABToIJ(int *i, unsigned int *j, const int64_t a, const uint64_t b, sieve_info_srcptr si);
-int ABTox(unsigned int *x, const int64_t a, const uint64_t b, sieve_info_srcptr si);
+int ABTox(uint64_t *x, const int64_t a, const uint64_t b, sieve_info_srcptr si);
 int ABToNx(unsigned int * N, unsigned int *x, const int64_t a, const uint64_t b, sieve_info_srcptr si);
 /*  */
 
 /* Warning: b might be negative, in which case we return (-a,-b) */
-static inline void xToAB(int64_t *a, uint64_t *b, const unsigned int x, sieve_info_srcptr si)
+static inline void xToAB(int64_t *a, uint64_t *b, const uint64_t x, sieve_info_srcptr si)
 {
     int i, j;
     int64_t c;
@@ -39,7 +39,7 @@ static inline void xToAB(int64_t *a, uint64_t *b, const unsigned int x, sieve_in
 }
 static inline void NxToAB(int64_t *a, uint64_t *b, const unsigned int N, const unsigned int x, sieve_info_srcptr si)
 {
-    xToAB(a, b, (N << LOG_BUCKET_REGION) + x, si);
+    xToAB(a, b, (((uint64_t)N) << LOG_BUCKET_REGION) + (uint64_t)x, si);
 }
 
 #endif	/* LAS_COORDINATES_H_ */
